Keep GetContentUrls inline in PlatformsBackOfficeController
Revert the mediator GetContentUrlsQuery/handler extraction; the URL-building
logic stays in the controller action. Retains the published-state gate and the
Permitted-free ContentUrlsRes.
