+++ src/Cloud/N3O.Umbraco.Cloud.Platforms/Controllers/PlatformsBackOfficeController.cs
@@ -1,20 +1,40 @@
+using Flurl;
 using Microsoft.AspNetCore.Mvc;
 using N3O.Umbraco.Attributes;
+using N3O.Umbraco.Cloud.Platforms.Extensions;
 using N3O.Umbraco.Cloud.Platforms.Models;
 using N3O.Umbraco.Cloud.Platforms.Queries;
+using N3O.Umbraco.Content;
+using N3O.Umbraco.Extensions;
 using N3O.Umbraco.Hosting;
 using N3O.Umbraco.Mediator;
+using N3O.Umbraco.Utilities;
+using Slugify;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Umbraco.Cms.Core.Services;
 
 namespace N3O.Umbraco.Cloud.Platforms.Controllers;
 
 [ApiDocument(PlatformsConstants.BackOfficeApiName)]
 public class PlatformsBackOfficeController : BackofficeAuthorizedApiController {
+    private readonly IContentCache _contentCache;
+    private readonly IContentService _contentService;
+    private readonly IContentTypeService _contentTypeService;
     private readonly IMediator _mediator;
+    private readonly ISlugHelper _slugHelper;
 
-    public PlatformsBackOfficeController(IMediator mediator) {
+    public PlatformsBackOfficeController(IContentCache contentCache,
+                                         IContentService contentService,
+                                         IContentTypeService contentTypeService,
+                                         IMediator mediator,
+                                         ISlugHelper slugHelper) {
+        _contentCache = contentCache;
+        _contentService = contentService;
+        _contentTypeService = contentTypeService;
         _mediator = mediator;
+        _slugHelper = slugHelper;
     }
 
     [HttpPost("previewHtml/{contentId:guid}")]
@@ -25,8 +45,38 @@ public class PlatformsBackOfficeController : BackofficeAuthorizedApiController {
     }
 
     [HttpGet("contentUrls/{contentId:guid}")]
-    public async Task<ActionResult<ContentUrlsRes>> GetContentUrls() {
-        var res = await _mediator.SendAsync<GetContentUrlsQuery, None, ContentUrlsRes>(None.Empty);
+    public async Task<ActionResult<ContentUrlsRes>> GetContentUrls([FromRoute] Guid contentId) {
+        var content = _contentService.GetById(contentId);
+        var urlSettings = _contentCache.Single<UrlSettingsContent>();
+
+        var isCampaign = content != null && content.IsCampaign(_contentTypeService);
+        var isOffering = content != null && content.IsOffering(_contentTypeService);
+
+        if (content == null || !content.Published || urlSettings == null || (!isCampaign && !isOffering)) {
+            return null;
+        }
+
+        string path;
+
+        if (isCampaign) {
+            path = _contentCache.GetCampaignPath(_slugHelper, content.Name);
+        } else {
+            var parent = _contentService.GetById(content.ParentId);
+
+            path = parent == null ? null : _contentCache.GetOfferingPath(_slugHelper, parent.Name, content.Name);
+        }
+
+        var res = new ContentUrlsRes();
+
+        if (path.HasValue()) {
+            if (urlSettings.StagingBaseUrl.HasValue()) {
+                res.StagingUrl = new Url(urlSettings.StagingBaseUrl).AppendPathSegment(path);
+            }
+
+            if (urlSettings.ProductionBaseUrl.HasValue()) {
+                res.ProductionUrl = new Url(urlSettings.ProductionBaseUrl).AppendPathSegment(path);
+            }
+        }
 
         return Ok(res);
     }
